Consolidate agent and reproduction helpers
- Inline single-consumer AgentComponents and FrameContext dataclasses

- Move asexual offspring creation into ReproductionService

- Route parent energy transfer through apply_energy_delta

Validation:

- .\.venv\Scripts\python.exe tools\smoke_gate.py

- .\.venv\Scripts\python.exe -m pytest tests\test_per_kind_policy_evolution.py tests\test_arch_energy_mutations.py

- .\.venv\Scripts\python.exe tools\fast_gate.py
